fix(scripts): symlink only the configs that exist

+++ scripts/unix/symlink-dotfiles.sh
@@ -4,11 +4,17 @@ DEFAULT_PROFILE="personal"
 
 cd $(dirname "$0")/../../profiles/"${1:-$DEFAULT_PROFILE}"/configurations
 
-rm ~/.bashrc
-ln -s "${PWD}"/bashrc ~/.bashrc
+if [ -f bashrc ]; then
+  rm ~/.bashrc
+  ln -s bashrc ~/.bashrc
+fi
 
-rm ~/.zshrc
-ln -s "${PWD}"/zshrc ~/.zshrc
+if [ -f zshrc ]; then
+  rm ~/.zshrc
+  ln -s zshrc ~/.zshrc
+fi
 
-rm ~/.vimrc
-ln -s "${PWD}"/vimrc ~/.vimrc
+if [ -f vimrc ]; then
+  rm ~/.vimrc
+  ln -s vimrc ~/.vimrc
+fi
